gozu_contable/requirements.txt
customtkinter>=5.2.0
tkcalendar>=1.6.1
pytesseract>=0.3.10
Pillow>=10.0.0
SpeechRecognition>=3.10.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
reportlab>=4.0.0
openpyxl>=3.1.0
tkinterdnd2>=0.3.0
